requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
aiohttp==3.11.11
Pillow==11.1.0
